EDT-765: Revision 580 allowed non-standard ontology items in ADL and XML, but only for term definitions. Any non-standard constraint definition items that may have existed in the archetype were lost when it was saved.
The non-standard items are now preserved when you save to ADL and XML. (Note, however, that there is no way to view or edit these non-standard constraint definition items. Perhaps we could do this in future on the Terminology tab.)

While doing this, many inconsistencies in the code between term definitions and constraint definitions were cleaned up. This is likely to fix other bugs. In particular, under certain circumstances the XML would be saved with empty comment tags; this is now fixed.