Suppress compiler warnings about the format of the version number. Treat all warnings as errors.
EDT-670: Because the latest ADL Parser DLL was built for 32-bit (not to run on any CPU), Archetype Editor has to be targeted to 32-bit mode so that it can run on 64-bit operating systems. We were achieving this via a post-build event that ran corflags, but this seems to be unnecessary: Visual Basic projects have an advanced setting that achieves the same thing.